app/requirements.txt
discord.py[voice]
ruamel.yaml
cassiopeia
